working webassembly init
