+++ website/index.html
@@ -71,6 +71,7 @@
         <div id="root">
             <div class="spinner">
                 <div class="dot1"></div>
+                <div class="dot2"></div>
             </div>
         </div>
         <script type="module" src="/src/index.jsx"></script>
