+++ website/src/workers/WorkerProvider.jsx
@@ -18,7 +18,6 @@ const WorkerProvider = ({ children }) => {
         };
 
         return () => {
-            console.log("Worker terminated");
             worker.terminate();
         };
     }, []);
@@ -28,7 +27,6 @@ const WorkerProvider = ({ children }) => {
             <>
                 <div className="spinner">
                     <div className="dot1"></div>
-                    <div className="dot2"></div>
                 </div>
             </>
         );
